requirements.txt
numpy
pandas
#pyaml
pycparse
python-dateutil
pytz
PyYAML
six
SoundFile
librosa
numba
tornado
tqdm
natsort
crepe
h5py
jupyter
jiwer
